feat: Moves sound limiter to separate class.
