+++ src/main/java/org/jitsi/jigasi/sounds/SoundNotificationManager.java
@@ -850,73 +850,6 @@ public class SoundNotificationManager
     }
 
     /**
-     * Implements RateLimiter for sound notifications.
-     */
-    private static class SoundRateLimiter
-    {
-        /**
-         * Initial time point.
-         */
-        private final AtomicReference<Instant> startTimePoint = new AtomicReference<>(null);
-
-        /**
-         * Timeout in milliseconds.
-         */
-        private long limiterTimeout;
-
-        /**
-         * SoundRateLimiter constructor.
-         *
-         * @param maxTimeout Timeout in milliseconds to block notification.
-         */
-        SoundRateLimiter(long maxTimeout)
-        {
-            this.limiterTimeout = maxTimeout;
-        }
-
-        /**
-         * Checks if timeout since last notification has passed.
-         * When it returns false, sound notification can be played.
-         *
-         * @return true of enabled, false otherwise.
-         */
-        public boolean on()
-        {
-            if (this.startTimePoint.compareAndSet(null, Instant.now()))
-            {
-                return false;
-            }
-
-            Instant prevTimePoint = this.startTimePoint.get();
-
-            if (prevTimePoint == null)
-            {
-                return false;
-            }
-
-            long elapsedMs =
-                Instant.now().toEpochMilli()
-                    - prevTimePoint.toEpochMilli();
-
-            if (elapsedMs >= this.limiterTimeout)
-            {
-                this.startTimePoint.set(Instant.now());
-                return false;
-            }
-
-            return true;
-        }
-
-        /**
-         * {@inheritDoc}
-         */
-        public void reset()
-        {
-            this.startTimePoint.set(null);
-        }
-    }
-
-    /**
      * Extracts MediaStream from call.
      * @param call the call.
      * @return null or <tt>MediaStream</tt> if available.
